refactor(types): drop manual MUI sync and obsolete version-alignment aliases
Post catalog sweep, react/MUI/emotion are uniform across apps and
frontend. That makes the manually-copied design-system type block in
apps and a few build-tool aliases redundant.

- Inline the MUI declare-module blocks (Button/IconButton/Breadcrumbs
  + ButtonOwnProps anchor attrs) in component-library's code.org
  theme entry. They now ride in dist/themes/code.org/index.d.ts.
  vite-plugin-dts strips side-effect imports without exported types,
  so a separate types.d.ts never reached consumers — top-level
  declare blocks in the entry survive emit.
- Delete the redundant src/themes/code.org/types.d.ts and the
  unreferenced package-root types/mui.d.ts (not in package.json
  exports).
- Drop the manually-copied Button/IconButton/Breadcrumbs blocks from
  apps/src/types/mui.d.ts; keep the apps-specific Typography
  augmentations. Design-system blocks flow in transitively via
  @code-dot-org/component-library/themes (already imported by
  apps/src/util/brand.ts and the studentSnapshot stories).
- Drop the @mui/material paths in apps/tsconfig.build.json. Normal
  node resolution lands on apps's MUI without help.
- Drop the react/react-dom aliases in frontend/apps/studio/vite.config.ts.
  yarn dedupe already eliminated nested copies in the frontend tree
  (verified: only frontend/node_modules/react exists).

Verified end-to-end: apps tsc clean over real call sites that depend
on the augmentation (e.g., size="extraSmall", color="white" on Button
in apps/src/aiDifferentiation/AiDiffBotMessageFooter.tsx). Stock MUI
types reject those values, so a passing typecheck proves the
component-library augmentations reach apps's TS program.

Aliases that stay (verified empirically against the live tree):
apps/webpack.config.js's react and @mui/material, apps/jest.config.js's
^react\$, apps's @types/react resolutions. component-library is
portal:-linked, so default node/webpack resolution from the symlinked
package reaches frontend/node_modules — two physical copies of
react@18.3.1 and @mui/material@7.3.2 coexist with !== singletons,
useTheme, ThemeProvider, and Button. The aliases force a single copy.
