--- frontend/packages/component-library/types/mui.d.ts
@@ -1,59 +1,0 @@
-declare module '@mui/material/styles' {
-  // Custom Typography definitions
-  interface TypographyVariants {
-    body3: React.CSSProperties;
-    body4: React.CSSProperties;
-    overline1: React.CSSProperties;
-    overline2: React.CSSProperties;
-    overline3: React.CSSProperties;
-    figcaption: React.CSSProperties;
-    label1: React.CSSProperties;
-    label2: React.CSSProperties;
-    label3: React.CSSProperties;
-    label4: React.CSSProperties;
-    strong: React.CSSProperties;
-    em: React.CSSProperties;
-  }
-
-  interface TypographyVariantsOptions {
-    body3?: React.CSSProperties;
-    body4?: React.CSSProperties;
-    overline1?: React.CSSProperties;
-    overline2?: React.CSSProperties;
-    overline3?: React.CSSProperties;
-    figcaption?: React.CSSProperties;
-    label1?: React.CSSProperties;
-    label2?: React.CSSProperties;
-    label3?: React.CSSProperties;
-    label4?: React.CSSProperties;
-    strong?: React.CSSProperties;
-    em?: React.CSSProperties;
-  }
-}
-
-declare module '@mui/material/Typography' {
-  interface TypographyPropsVariantOverrides {
-    body3: true;
-    body4: true;
-    overline: false; // disable the default overline
-    overline1: true; // add overline1 variant to match DSCO naming pattern
-    overline2: true;
-    overline3: true;
-    caption: false; // disable the default caption
-    figcaption: true; // add figcaption variant to match DSCO naming pattern
-    label1: true;
-    label2: true;
-    label3: true;
-    label4: true;
-    strong: true;
-    em: true;
-  }
-}
-
-declare module '@mui/material/Breadcrumbs' {
-  interface BreadcrumbsProps {
-    size?: 'xs' | 's' | 'm' | 'l';
-  }
-}
-
-export {};
